+++ components/dynamic-menu/__docs__/storybook-stories.jsx
@@ -1,13 +1,13 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import IconSettings from '../../icon-settings';
-
-// import DynamicMenu from '../../dynamic-menu';
-import { DYNAMIC_MENU } from '../../../utilities/constants';
-
-
-import DynamicMenu10ItemsExample from '../__examples__/dynamic-menu-10-items';
-
-storiesOf(DYNAMIC_MENU, module)
-	.addDecorator((getStory) => <div className="slds-p-around--medium"><IconSettings iconPath="/assets/icons">{getStory()}</IconSettings></div>)
-	.add('Base', () => (<DynamicMenu10ItemsExample />));
+// import React from 'react';
+// import { storiesOf } from '@storybook/react';
+// import IconSettings from '../../icon-settings';
+//
+// // import DynamicMenu from '../../dynamic-menu';
+// import { DYNAMIC_MENU } from '../../../utilities/constants';
+//
+//
+// import DynamicMenu10ItemsExample from '../__examples__/dynamic-menu-10-items';
+//
+// storiesOf(DYNAMIC_MENU, module)
+// 	.addDecorator((getStory) => <div className="slds-p-around--medium"><IconSettings iconPath="/assets/icons">{getStory()}</IconSettings></div>)
+// 	.add('Base', () => (<DynamicMenu10ItemsExample />));
