+++ components/storybook-stories.js
@@ -11,7 +11,6 @@ export ButtonStateful from '../components/button-stateful/__docs__/storybook-sto
 export Card from '../components/card/__docs__/storybook-stories';
 export Checkbox from '../components/forms/checkbox/__docs__/storybook-stories';
 export Combobox from '../components/combobox/__docs__/storybook-stories';
-export DynamicMenu from '../components/dynamic-menu/__docs__/storybook-stories';
 export Filter from '../components/filter/__docs__/storybook-stories';
 export GlobalNavigationBar from '../components/global-navigation-bar/__docs__/storybook-stories';
 export DataTable from '../components/data-table/__docs__/storybook-stories';
